# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export type RegistrationInputAuthorityType = typeof RegistrationInputAuthorityType[keyof typeof RegistrationInputAuthorityType];


export const RegistrationInputAuthorityType = {
  kitchen: 'kitchen',
  fpu: 'fpu',
} as const;

export interface RegistrationInput {
  /** @minLength 2 */
  organizationName: string;
  /** @minLength 2 */
  representative: string;
  /** @minLength 7 */
  phone: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 4 */
  address: string;
  /** @minLength 2 */
  location: string;
  /** @minLength 2 */
  proofId: string;
  authorityType: RegistrationInputAuthorityType;
  /** @nullable */
  fpuType?: string | null;
  /** @nullable */
  productCategory?: string | null;
}

export interface Registration {
  id: number;
  organizationName: string;
  authorityType: string;
  status: string;
  message: string;
}

export interface OtpInput {
  /**
     * @minLength 4
     * @maxLength 6
     */
  otp: string;
}

export interface Application {
  id: number;
  organizationName: string;
  representative: string;
  contact: string;
  location: string;
  authorityType: string;
  status: string;
  submittedAt: string;
  proofId?: string;
}

export type DecisionInputDecision = typeof DecisionInputDecision[keyof typeof DecisionInputDecision];


export const DecisionInputDecision = {
  approve: 'approve',
  reject: 'reject',
  more_info: 'more_info',
} as const;

export interface DecisionInput {
  decision: DecisionInputDecision;
  note?: string;
}

export interface Metric {
  label: string;
  value: number;
  unit: string;
  trend: number;
}

export interface Activity {
  title: string;
  detail: string;
  time: string;
  kind: string;
}

export interface Dashboard {
  organizationName: string;
  role: string;
  status: string;
  points: number;
  leaderboard: number;
  stats: Metric[];
  activities: Activity[];
}

export type PlanInputMode = typeof PlanInputMode[keyof typeof PlanInputMode];


export const PlanInputMode = {
  kitchen: 'kitchen',
  fpu: 'fpu',
} as const;

export interface PlanInput {
  mode: PlanInputMode;
  /** @minLength 2 */
  product: string;
  date: string;
  /** @minimum 1 */
  quantity: number;
  /** @minimum 0 */
  currentStock?: number;
  /** @minimum 0 */
  confirmedOrders?: number;
  meal?: string;
  events?: string;
  constraints?: string;
}

export interface PlanResult {
  product: string;
  expectedDemand: number;
  requiredProduction: number;
  safetyBuffer: number;
  recommended: number;
  explanation: string;
  history: Metric[];
}

export interface EnergyInput {
  recipe: string;
  /** @minimum 1 */
  quantity: number;
  equipment: string;
  source: string;
}

export interface EnergyResult {
  method: string;
  equipment: string;
  estimatedEnergy: number;
  savings: number;
  co2Reduction: number;
}

export interface InventoryInput {
  product: string;
  /** @minimum 1 */
  quantity: number;
}

export interface InventoryLine {
  ingredient: string;
  required: number;
  available: number;
  shortage: number;
  surplus: number;
}

export interface RecoveryInput {
  material: string;
  /** @minimum 0 */
  quantity: number;
  category: string;
  condition: string;
  location: string;
}

export interface RecoveryListing {
  id: number;
  material: string;
  quantity: number;
  category: string;
  condition: string;
  location: string;
  availability: string;
  matchCount: number;
}

export interface StorageReading {
  room: string;
  item: string;
  quantity: number;
  temperature: number;
  humidity: number;
  timestamp: string;
  status: string;
  limit: number;
}

export interface MachineReading {
  machine: string;
  status: string;
  power: number;
  energy: number;
  runtime: number;
  downtime: number;
  output: number;
  events: string[];
}

export interface ImpactPoint {
  label: string;
  value: number;
}

export interface Impact {
  foodRescued: number;
  foodDonated: number;
  deliveries: number;
  wastePrevented: number;
  energySaved: number;
  co2Avoided: number;
  points: number;
  leaderboard: number;
  series: ImpactPoint[];
}

export type GetDashboardParams = {
role: GetDashboardRole;
organization?: string;
};

export type GetDashboardRole = typeof GetDashboardRole[keyof typeof GetDashboardRole];


export const GetDashboardRole = {
  kitchen: 'kitchen',
  fpu: 'fpu',
  receiver: 'receiver',
  admin: 'admin',
} as const;

export type GetImpactParams = {
range?: GetImpactRange;
};

export type GetImpactRange = typeof GetImpactRange[keyof typeof GetImpactRange];


export const GetImpactRange = {
  today: 'today',
  month: 'month',
  year: 'year',
} as const;