# lib/api-zod/src/generated/types/dashboard.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Activity } from './activity';
import type { Metric } from './metric';

export interface Dashboard {
  organizationName: string;
  role: string;
  status: string;
  points: number;
  leaderboard: number;
  stats: Metric[];
  activities: Activity[];
}