# lib/api-zod/src/generated/types/decisionInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */
import type { DecisionInputDecision } from './decisionInputDecision';

export interface DecisionInput {
  decision: DecisionInputDecision;
  note?: string;
}