# lib/api-zod/src/generated/types/energyResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */

export interface EnergyResult {
  method: string;
  equipment: string;
  estimatedEnergy: number;
  savings: number;
  co2Reduction: number;
}