# lib/api-zod/src/generated/types/machineReading.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */

export interface MachineReading {
  machine: string;
  status: string;
  power: number;
  energy: number;
  runtime: number;
  downtime: number;
  output: number;
  events: string[];
}