# lib/api-zod/src/generated/types/otpInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */

export interface OtpInput {
  /**
     * @minLength 4
     * @maxLength 6
     */
  otp: string;
}