# lib/api-zod/src/generated/types/planResult.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */
import type { Metric } from './metric';

export interface PlanResult {
  product: string;
  expectedDemand: number;
  requiredProduction: number;
  safetyBuffer: number;
  recommended: number;
  explanation: string;
  history: Metric[];
}