# lib/api-zod/src/generated/types/recoveryListing.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */

export interface RecoveryListing {
  id: number;
  material: string;
  quantity: number;
  category: string;
  condition: string;
  location: string;
  availability: string;
  matchCount: number;
}