# lib/api-zod/src/generated/types/registration.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */

export interface Registration {
  id: number;
  organizationName: string;
  authorityType: string;
  status: string;
  message: string;
}