# lib/api-zod/src/generated/types/registrationInput.ts
/**
 * Generated by orval 🍺
 * Do not edit manually.
 * Api
 * Waste2Value interactive sustainability platform API
 * OpenAPI spec version: 0.1.0
 */
import type { RegistrationInputAuthorityType } from './registrationInputAuthorityType';

export interface RegistrationInput {
  /** @minLength 2 */
  organizationName: string;
  /** @minLength 2 */
  representative: string;
  /** @minLength 7 */
  phone: string;
  email: string;
  /** @minLength 6 */
  password: string;
  /** @minLength 4 */
  address: string;
  /** @minLength 2 */
  location: string;
  /** @minLength 2 */
  proofId: string;
  authorityType: RegistrationInputAuthorityType;
  /** @nullable */
  fpuType?: string | null;
  /** @nullable */
  productCategory?: string | null;
}